changed timestamp to Long

+++ src/main/kotlin/marudor/hafas/IRISMessage.kt
@@ -2,7 +2,7 @@ package marudor.hafas
 
 data class IRISMessage(
     var text: String? = null,
-    var timestamp: Int? = null,
+    var timestamp: Long? = null,
     var superseded: Boolean? = null,
     var priority: Int? = null
 ) {
